requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pandas>=2.2.2
openpyxl>=3.1.4
python-multipart>=0.0.9
pydantic>=2.7.4
numpy>=1.26.4
PyJWT
passlib[bcrypt]
bcrypt
